refactor(fizruk): move Today's plan & recovery atlas from Dashboard to Plan page, drop PhotoProgress from Body
